refactor: move Routes to constants
